replace Clients services providers with interfaces
